requirements.txt
Django>=4.0
djangorestframework
drf-yasg[validation]
django-filter
psycopg2-binary
gunicorn
flake8
pytest
pytest-django
black
